Moved rememberEventListener into CircuitContent where it is used
